Move event actions to subdirectory
